move initilize code to monitor
